add tools

+++ src/components/ui/slider.tsx
@@ -15,7 +15,7 @@ interface SliderProps {
 const Slider = React.forwardRef<HTMLInputElement, SliderProps>(
   ({ value, onValueChange, max = 100, min = 0, step = 1, className, ...props }, ref) => {
     const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-      onValueChange([parseInt(e.target.value)])
+      onValueChange([parseFloat(e.target.value)])
     }
 
     return (
